chore(nest): move test files to correct directory structure
